==> Manuscript/MS_tables_figs_rev1/2_make_CJFAS_figs_rev1.R <==
# 2_make_CJFAS_figs_rev1.R
# exploring alternative figures for CJFAS based on first round of reviewer comments

library(ggplot2); library(gridExtra); library(grid); library(ggpubr)
library(dplyr); library(tidyr)

# used for Windows 
windowsFonts(Times=windowsFont("Times New Roman"))
# end used for Windows section

# set working directory to source file location to begin

# specify order for IBMs so Ensemble and SCAA are at the end
IBM_order <- rev(c("AIM","CC-FM","CC-FSPR","DLM","ES-FM","ES-Frecent",
               "ES-FSPR","ES-Fstable","Islope","Ismooth","Itarget",
               "Skate","Ensemble","SCAA"))
IBM_order_new <- rev(c("AIM","CC-FM","CC-FSPR","DynLin","ES-FM","ES-Frecent",
                       "ES-FSPR","ES-Fstable","Islope","Ismooth","Itarget",
                       "Skate","Ensemble","SCAA"))

# get data for original figures
all_scen <- read.csv("../MS_tables_figs/PMs_all_scenarios_new.csv",
                     header=TRUE,stringsAsFactors=FALSE) 
retro_Fhist <- read.csv("../MS_tables_figs/PMs_by_retro_Fhistory_new.csv",
                        header=TRUE,stringsAsFactors=FALSE)
time_Fhist <- read.csv("../MS_tables_figs/PMs_by_time_Fhist_new.csv",
                    header=TRUE,stringsAsFactors=FALSE)
Cmult_Fhist <- read.csv("../MS_tables_figs/PMs_by_Cmult_Fhist_new.csv",
                        header=TRUE,stringsAsFactors=FALSE)
Cmult_Fhist$Catch_mult <- factor(Cmult_Fhist$Catch_mult)

ssb_median_by_scenario <- readRDS("../tables_figs/ssb_median_by_scenario.rds") %>%
  mutate(metric = paste0("ssb_", metric))

f_median_by_scenario <- readRDS("../tables_figs/f_median_by_scenario.rds") %>%
  mutate(metric = paste0("f_", metric))

catch_median_by_scenario <- readRDS("../tables_figs/catch_median_by_scenario.rds") %>%
  mutate(metric = paste0("catch_", metric))

td_med <- rbind(ssb_median_by_scenario, f_median_by_scenario, catch_median_by_scenario)

td3_l_med <- td_med %>%
  filter(metric %in% c("ssb_l_avg_ssb_ssbmsy", "catch_l_avg_catch_msy")) %>%
  distinct() %>%
  pivot_wider(names_from = metric, values_from = value) %>%
  rename(x_value = ssb_l_avg_ssb_ssbmsy, y_value = catch_l_avg_catch_msy)

sims5 <- readRDS("../MS_tables_figs/Fig5_sims.rds")

# get data for new and alternative figures
sims <- readRDS("sims.RDS")

# some data wrangling for new and alternative figures
# compute medians by IBMs
medIBM <- sims %>%
  group_by(IBMlab, metric) %>%
  summarize(medval = median(value))

# compute medians by IBMs and Scenarios
medIBMScen <- sims %>%
  group_by(IBMlab, Scenlab, metric) %>%
  summarize(medval = median(value))

# get long term SSB/SSBmsy, F/Fmsy, and C/MSY medians
medlong <- sims %>%
  filter(time.avg == "L",
         metric %in% c("avg_ssb_ssbmsy", "avg_f_fmsy", "avg_catch_msy")) %>%
  group_by(IBMlab, Scenlab, metric) %>%
  summarize(medval = median(value)) %>%
  pivot_wider(names_from = metric, values_from = medval) %>%
  mutate(retro_type = ifelse(substr(Scenlab, 1, 1)=="C", "Catch", "M"))

# medians by retro type and F history
medretro_Fhist <- sims %>%
  group_by(IBMlab, retro_type, Fhist, metric) %>%
  summarize(medval = median(value)) %>%
  mutate(F_history = ifelse(Fhist==1, "Fmsy", "2.5 x Fmsy"),
         Retro_source = retro_type)

# medians by time period and F history for F/Fmsy only
medtime_Fhist <- sims %>%
  filter(metric == "avg_f_fmsy") %>%
  group_by(IBMlab, time.avg, Fhist, metric) %>%
  summarize(medval = median(value)) %>%
  mutate(F_history = ifelse(Fhist==1, "Fmsy", "2.5 x Fmsy"),
         Period = ifelse(time.avg=="L", "Long-term", ifelse(time.avg=="S","Short-term","All years or combined long and short")))

# medians by catch multiplier and F history for F/Fmsy only
medCmult_Fhist <- sims %>%
  filter(metric == "avg_f_fmsy") %>%
  group_by(IBMlab, catch.mult, Fhist, metric) %>%
  summarize(medval = median(value)) %>%
  mutate(Catch_mult = factor(catch.mult, level=c("0.75", "1")),
         F_history = ifelse(Fhist==1, "Fmsy", "2.5 x Fmsy"))

# 1000 point plots for long-term SSB/SSBmsy, F/Fmsy, and Catch/MSY  for Scenario CF1A only
simsfig5 <- sims %>%
  filter(metric %in% c("avg_ssb_ssbmsy","avg_f_fmsy","avg_catch_msy"),
         time.avg == "L",
         Scenlab == "CF1A") %>%
  select(isim, IBMlab, metric, value) %>%
  pivot_wider(names_from = metric, values_from = value)


# Figure labels for each PM
SSB_label <-expression(paste("SSB / ",SSB[MSY],sep = ""))
F_label <-expression(paste("F / ",F[MSY],sep = ""))
C_label <- "C / MSY"
IAV_label <- "Relative variability in catch (IAV)"
PING_label <- "Probability of overfishing"           # orig figs only
PED_label <- "Probability of being overfished"       # orig figs only
NING_label <- "Number of years overfishing"          # new figs only
NED_label <- "Number of years overfished"            # new figs only

# colors used to distinguish two types of points
point_cols <- c("black","dark gray")

# each figure can be generated by running the open curly bracket at the start of the fig
# set up to run on Windows currently, add quartz() on first line after curly bracket to run on Mac

# Figure 1 ---------------------------------------------------------------
# Original submission
{
  #quartz() # can only run this on Mac
  pSSB<-ggplot(data=all_scen, aes(x=factor(IBM,level=IBM_order), y=SSB)) +
  geom_bar(stat="identity",fill="gray")+
  theme_classic() + 
  theme(text=element_text(family="Times"))+
  xlab("") + ylab(SSB_label) + ggtitle("A")+
  geom_hline(yintercept = 1) +
  scale_y_continuous(expand = c(0,0))+ 
  coord_flip()

pF<-ggplot(data=all_scen, aes(x=factor(IBM,level=IBM_order), y=F)) +
  geom_bar(stat="identity",fill="gray")+
  theme_classic() + 
  theme(text=element_text(family="Times"))+
  xlab("") + ylab(F_label) + ggtitle("B")+
  geom_hline(yintercept = 1) +
  scale_y_continuous(expand = c(0,0))+ 
  coord_flip()

pC<-ggplot(data=all_scen, aes(x=factor(IBM,level=IBM_order), y=Catch)) +
  geom_bar(stat="identity",fill="gray")+
  theme_classic() + 
  theme(text=element_text(family="Times"))+
  xlab("") + ylab(C_label) + ggtitle("C")+
  geom_hline(yintercept = 1) +
  scale_y_continuous(expand = c(0,0))+ 
  coord_flip()

pIAV<-ggplot(data=all_scen, aes(x=factor(IBM,level=IBM_order), y=IAV)) +
  geom_bar(stat="identity",fill="gray")+
  theme_classic() + 
  theme(text=element_text(family="Times"))+
  xlab("") + ylab(IAV_label) + ggtitle("D") +
  scale_y_continuous(expand = c(0,0))+ 
  coord_flip()

pOFG<-ggplot(data=all_scen, aes(x=factor(IBM,level=IBM_order), y=PING)) +
  geom_bar(stat="identity",fill="gray")+
  theme_classic() + 
  theme(text=element_text(family="Times"))+ 
  xlab("") + ylab(PING_label) + ggtitle("E") +
  scale_y_continuous(expand = c(0,0))+
  geom_hline(yintercept = 0.5)+ 
  coord_flip()

pOFD<-ggplot(data=all_scen, aes(x=factor(IBM,level=IBM_order), y=PED)) +
  geom_bar(stat="identity",fill="gray")+
  theme_classic() + 
  theme(text=element_text(family="Times"))+
  xlab("") + ylab(PED_label) + ggtitle("F") +
  scale_y_continuous(expand = c(0,0))+ 
  coord_flip()

p1 <- grid.arrange(pSSB,pF,pC,pIAV,pOFG,pOFD,nrow=3)
ggsave("Figure_1_orig.pdf",plot=p1,device="pdf",width=7,height=8)
}

# new version of Figure 1 using number of years overfishing and overfished
{
pSSBn<-ggplot(data=filter(medIBM, metric=="avg_ssb_ssbmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=medval)) +
    geom_bar(stat="identity",fill="gray")+
    theme_classic() + 
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(SSB_label) + ggtitle("A")+
    geom_hline(yintercept = 1) +
    scale_y_continuous(expand = c(0,0))+ 
    coord_flip()
  
pFn<-ggplot(data=filter(medIBM, metric=="avg_f_fmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=medval)) +
    geom_bar(stat="identity",fill="gray")+
    theme_classic() + 
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(F_label) + ggtitle("B")+
    geom_hline(yintercept = 1) +
    scale_y_continuous(expand = c(0,0))+ 
    coord_flip()
  
pCn<-ggplot(data=filter(medIBM, metric=="avg_catch_msy"), aes(x=factor(IBMlab,level=IBM_order_new), y=medval)) +
    geom_bar(stat="identity",fill="gray")+
    theme_classic() + 
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(C_label) + ggtitle("C")+
    geom_hline(yintercept = 1) +
    scale_y_continuous(expand = c(0,0))+ 
    coord_flip()
  
pIAVn<-ggplot(data=filter(medIBM, metric=="iav_catch"), aes(x=factor(IBMlab,level=IBM_order_new), y=medval)) +
    geom_bar(stat="identity",fill="gray")+
    theme_classic() + 
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(IAV_label) + ggtitle("D") +
    scale_y_continuous(expand = c(0,0))+ 
    coord_flip()
  
nINGn<-ggplot(data=filter(medIBM, metric=="n_gr_fmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=medval)) +
    geom_bar(stat="identity",fill="gray")+
    theme_classic() + 
    theme(text=element_text(family="Times"))+ 
    xlab("") + ylab(NING_label) + ggtitle("E") +
    scale_y_continuous(expand = c(0,0))+
    coord_flip()
  
nEDn<-ggplot(data=filter(medIBM, metric=="n_less_05_bmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=medval)) +
    geom_bar(stat="identity",fill="gray")+
    theme_classic() + 
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(NED_label) + ggtitle("F") +
    scale_y_continuous(expand = c(0,0))+ 
    coord_flip()
  
p1n <- grid.arrange(pSSBn,pFn,pCn,pIAVn,nINGn,nEDn,nrow=3)
ggsave("Figure_1_new.pdf",plot=p1n,device="pdf",width=7,height=8)
}

# Alternative 1 for Figure 1
# add points showing medians for each of the 16 scenarios
{
pSSBalt1 <- pSSBn +
  geom_point(data=filter(medIBMScen, metric=="avg_ssb_ssbmsy"), aes(x=IBMlab, y=medval))
pFalt1 <- pFn +
  geom_point(data=filter(medIBMScen, metric=="avg_f_fmsy"), aes(x=IBMlab, y=medval))
pCalt1 <- pCn +
  geom_point(data=filter(medIBMScen, metric=="avg_catch_msy"), aes(x=IBMlab, y=medval))
pIAValt1 <- pIAVn +
  geom_point(data=filter(medIBMScen, metric=="iav_catch"), aes(x=IBMlab, y=medval))
nINGalt1 <- nINGn +
  geom_point(data=filter(medIBMScen, metric=="n_gr_fmsy"), aes(x=IBMlab, y=medval))
nEDalt1 <- nEDn +
  geom_point(data=filter(medIBMScen, metric=="n_less_05_bmsy"), aes(x=IBMlab, y=medval))  
p1alt1 <- grid.arrange(pSSBalt1,pFalt1,pCalt1,pIAValt1,nINGalt1,nEDalt1,nrow=3)
ggsave("Figure_1_alt1.pdf",plot=p1alt1,device="pdf",width=7,height=8)
}

# Alternative 2 using violin plots
# could try to limit upper values of each plot to see the broader part of the violin better
{
pSSBalt2 <- ggplot(data=filter(sims, metric=="avg_ssb_ssbmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=value)) +
    geom_violin(draw_quantiles = c(0.25, 0.5, 0.75)) +
    theme_classic() +
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(SSB_label) + ggtitle("A")+
    geom_hline(yintercept = 1) +
    scale_y_continuous(expand = c(0,0))+ 
    coord_flip()
pFalt2 <- ggplot(data=filter(sims, metric=="avg_f_fmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=value)) +
  geom_violin(draw_quantiles = c(0.25, 0.5, 0.75)) +
  theme_classic() +
  theme(text=element_text(family="Times"))+
  xlab("") + ylab(F_label) + ggtitle("B")+
  geom_hline(yintercept = 1) +
  scale_y_continuous(expand = c(0,0))+ 
  coord_flip()
pCalt2 <- ggplot(data=filter(sims, metric=="avg_catch_msy"), aes(x=factor(IBMlab,level=IBM_order_new), y=value)) +
  geom_violin(draw_quantiles = c(0.25, 0.5, 0.75)) +
  theme_classic() +
  theme(text=element_text(family="Times"))+
  xlab("") + ylab(C_label) + ggtitle("C")+
  geom_hline(yintercept = 1) +
  scale_y_continuous(expand = c(0,0))+ 
  coord_flip()
pIAValt2 <- ggplot(data=filter(sims, metric=="iav_catch"), aes(x=factor(IBMlab,level=IBM_order_new), y=value)) +
  geom_violin(draw_quantiles = c(0.25, 0.5, 0.75)) +
  theme_classic() +
  theme(text=element_text(family="Times"))+
  xlab("") + ylab(IAV_label) + ggtitle("D")+
  scale_y_continuous(expand = c(0,0))+ 
  coord_flip()
pINGalt2 <- ggplot(data=filter(sims, metric=="n_gr_fmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=value)) +
  geom_violin(draw_quantiles = c(0.25, 0.5, 0.75)) +
  theme_classic() +
  theme(text=element_text(family="Times"))+
  xlab("") + ylab(NING_label) + ggtitle("E")+
  scale_y_continuous(expand = c(0,0))+ 
  coord_flip()
pEDalt2 <- ggplot(data=filter(sims, metric=="n_less_05_bmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=value)) +
  geom_violin(draw_quantiles = c(0.25, 0.5, 0.75)) +
  theme_classic() +
  theme(text=element_text(family="Times"))+
  xlab("") + ylab(NED_label) + ggtitle("F")+
  scale_y_continuous(expand = c(0,0))+ 
  coord_flip()
p1alt2 <- grid.arrange(pSSBalt2,pFalt2,pCalt2,pIAValt2,pINGalt2,pEDalt2,nrow=3)
ggsave("Figure_1_alt2.pdf",plot=p1alt2,device="pdf",width=7,height=8)
}

# Alternative 3 with boxplots
{
  pSSBalt3 <- ggplot(data=filter(sims, metric=="avg_ssb_ssbmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=value)) +
    geom_boxplot() +
    theme_classic() +
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(SSB_label) + ggtitle("A")+
    geom_hline(yintercept = 1) +
    scale_y_continuous(expand = c(0,0))+ 
    coord_flip()
  pFalt3 <- ggplot(data=filter(sims, metric=="avg_f_fmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=value)) +
    geom_boxplot() +
    theme_classic() +
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(F_label) + ggtitle("B")+
    geom_hline(yintercept = 1) +
    scale_y_continuous(expand = c(0,0))+ 
    coord_flip()
  pCalt3 <- ggplot(data=filter(sims, metric=="avg_catch_msy"), aes(x=factor(IBMlab,level=IBM_order_new), y=value)) +
    geom_boxplot() +
    theme_classic() +
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(C_label) + ggtitle("C")+
    geom_hline(yintercept = 1) +
    scale_y_continuous(expand = c(0,0))+ 
    coord_flip()
  pIAValt3 <- ggplot(data=filter(sims, metric=="iav_catch"), aes(x=factor(IBMlab,level=IBM_order_new), y=value)) +
    geom_boxplot() +
    theme_classic() +
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(IAV_label) + ggtitle("D")+
    scale_y_continuous(expand = c(0,0))+ 
    coord_flip()
  pINGalt3 <- ggplot(data=filter(sims, metric=="n_gr_fmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=value)) +
    geom_boxplot() +
    theme_classic() +
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(NING_label) + ggtitle("E")+
    scale_y_continuous(expand = c(0,0))+ 
    coord_flip()
  pEDalt3 <- ggplot(data=filter(sims, metric=="n_less_05_bmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=value)) +
    geom_boxplot() +
    theme_classic() +
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(NED_label) + ggtitle("F")+
    scale_y_continuous(expand = c(0,0))+ 
    coord_flip()
  p1alt3 <- grid.arrange(pSSBalt3,pFalt3,pCalt3,pIAValt3,pINGalt3,pEDalt3,nrow=3)
  ggsave("Figure_1_alt3.pdf",plot=p1alt3,device="pdf",width=7,height=8)
}

# Alternative 4 boxplots showing only interquartile range and median
# have to adjust the upper range of values by hand, need to figure out a better way to do this if want to go with it
{
  pSSBalt4 <- ggplot(data=filter(sims, metric=="avg_ssb_ssbmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=value)) +
    geom_boxplot(outlier.shape = NA, coef = 0, fill="light gray") +
    theme_classic() +
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(SSB_label) + ggtitle("A")+
    geom_hline(yintercept = 1) +
    coord_flip(ylim = c(0,3.5), expand=FALSE) 
  pFalt4 <- ggplot(data=filter(sims, metric=="avg_f_fmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=value)) +
    geom_boxplot(outlier.shape = NA, coef = 0, fill="light gray") +
    theme_classic() +
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(F_label) + ggtitle("B")+
    geom_hline(yintercept = 1) +
    coord_flip(ylim = c(0,9), expand=FALSE) 
  pCalt4 <- ggplot(data=filter(sims, metric=="avg_catch_msy"), aes(x=factor(IBMlab,level=IBM_order_new), y=value)) +
    geom_boxplot(outlier.shape = NA, coef = 0, fill="light gray") +
    theme_classic() +
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(C_label) + ggtitle("C")+
    geom_hline(yintercept = 1) +
    coord_flip(ylim = c(0,2), expand=FALSE) 
  pIAValt4 <- ggplot(data=filter(sims, metric=="iav_catch"), aes(x=factor(IBMlab,level=IBM_order_new), y=value)) +
    geom_boxplot(outlier.shape = NA, coef = 0, fill="light gray") +
    theme_classic() +
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(IAV_label) + ggtitle("D")+
    coord_flip(ylim = c(0,1), expand=FALSE) 
  pINGalt4 <- ggplot(data=filter(sims, metric=="n_gr_fmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=value)) +
    geom_boxplot(outlier.shape = NA, coef = 0, fill="light gray") +
    theme_classic() +
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(NING_label) + ggtitle("E")+
    coord_flip(ylim = c(0,26), expand=FALSE) 
  pEDalt4 <- ggplot(data=filter(sims, metric=="n_less_05_bmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=value)) +
    geom_boxplot(outlier.shape = NA, coef = 0, fill="light gray") +
    theme_classic() +
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(NED_label) + ggtitle("F")+
    coord_flip(ylim = c(0,26), expand=FALSE) 
  p1alt4 <- grid.arrange(pSSBalt4,pFalt4,pCalt4,pIAValt4,pINGalt4,pEDalt4,nrow=3)
  ggsave("Figure_1_alt4.pdf",plot=p1alt4,device="pdf",width=7,height=8)
}

# Figure 2 --------------------------------------------------------------- 
# Median long term C/MSY vs SSB/SSBmsy
# original
{
  fig2plot <- ggplot(td3_l_med, aes(x=x_value, y=y_value, color=retro_type)) +
    geom_point() +
    facet_wrap(~factor(IBMlab, level=rev(IBM_order))) +
    theme_classic() + 
    theme(text = element_text(family = "Times")) +
    labs(x=SSB_label, y=C_label) +
    scale_color_manual(values=point_cols)
  ggsave(filename = "Figure_2_orig.pdf", plot=fig2plot, device="pdf",width=6,height=6)
}

# with new data and DynLin label
{
  fig2new <- ggplot(medlong, aes(x=avg_ssb_ssbmsy, y=avg_catch_msy, color=retro_type)) +
    geom_point() +
    facet_wrap(~factor(IBMlab, level=rev(IBM_order_new))) +
    theme_classic() + 
    theme(text = element_text(family = "Times")) +
    labs(x=SSB_label, y=C_label) +
    scale_color_manual(values=point_cols)
  ggsave(filename = "Figure_2_new.pdf", plot=fig2new, device="pdf",width=6,height=6)
}

# similar plot with F/Fmsy on y-axis
{
  fig2alt1 <- ggplot(medlong, aes(x=avg_ssb_ssbmsy, y=avg_f_fmsy, color=retro_type)) +
    geom_point() +
    facet_wrap(~factor(IBMlab, level=rev(IBM_order_new))) +
    theme_classic() + 
    theme(text = element_text(family = "Times")) +
    labs(x=SSB_label, y=F_label) +
    scale_color_manual(values=point_cols)
  ggsave(filename = "Figure_2_alt1.pdf", plot=fig2alt1, device="pdf",width=6,height=6)
}

# similar plot with F/Fmsy on x-axis and C/MSY on y-axis
{
  fig2alt2 <- ggplot(medlong, aes(x=avg_f_fmsy, y=avg_catch_msy, color=retro_type)) +
    geom_point() +
    facet_wrap(~factor(IBMlab, level=rev(IBM_order_new))) +
    theme_classic() + 
    theme(text = element_text(family = "Times")) +
    labs(x=F_label, y=C_label) +
    scale_color_manual(values=point_cols)
  ggsave(filename = "Figure_2_alt2.pdf", plot=fig2alt2, device="pdf",width=6,height=6)
}

# Figure 3 ---------------------------------------------------------------
# PMs by F history and Retro source
# something wrong with the bottom two plots - not matching the fig in the submitted MS
{
#quartz() # can only run this on Mac
pSSB_retro<-ggplot(data=retro_Fhist, aes(x=factor(IBM,level=IBM_order), y=SSB)) +
  geom_point((aes(colour = Retro_source, shape=F_history)))+
  theme_classic() + 
  theme(text=element_text(family="Times"))+
  xlab("") + ylab(SSB_label) + ggtitle("A")+
  geom_hline(yintercept = 1) +
  scale_color_manual(values=point_cols)+
  coord_flip()

pF_retro<-ggplot(data=retro_Fhist, aes(x=factor(IBM,level=IBM_order), y=F)) +
  geom_point((aes(colour = Retro_source, shape=F_history)))+
  theme_classic() + 
  theme(text=element_text(family="Times"))+
  xlab("") + ylab(F_label) + ggtitle("B")+
  geom_hline(yintercept = 1) +
  scale_color_manual(values=point_cols)+
  coord_flip()

pC_retro<-ggplot(data=retro_Fhist, aes(x=factor(IBM,level=IBM_order), y=Catch)) +
  geom_point((aes(colour = Retro_source, shape=F_history)))+
  theme_classic() + 
  theme(text=element_text(family="Times"))+
  xlab("") + ylab(C_label) + ggtitle("C")+
  geom_hline(yintercept = 1) +
  scale_color_manual(values=point_cols)+
  coord_flip()

pIAV_retro<-ggplot(data=retro_Fhist, aes(x=factor(IBM,level=IBM_order), y=IAV)) +
  geom_point((aes(colour = Retro_source, shape=F_history)))+
  theme_classic() + 
  theme(text=element_text(family="Times"))+
  xlab("") + ylab(IAV_label) + ggtitle("D")+
  scale_color_manual(values=point_cols)+
  coord_flip()

pOFG_retro<-ggplot(data=retro_Fhist, aes(x=factor(IBM,level=IBM_order), y=PING)) +
  geom_point((aes(colour = Retro_source, shape=F_history)))+
  theme_classic() + 
  theme(text=element_text(family="Times"))+
  xlab("") + ylab(PING_label) + ggtitle("E")+
  geom_hline(yintercept = 0.5) +
  scale_color_manual(values=point_cols)+
  coord_flip()

pOFD_retro<-ggplot(data=retro_Fhist, aes(x=factor(IBM,level=IBM_order), y=PED)) +
  geom_point((aes(colour = Retro_source, shape=F_history)))+
  theme_classic() + 
  theme(text=element_text(family="Times"))+
  xlab("") + ylab(PED_label) + ggtitle("F")+
  scale_color_manual(values=point_cols)+
  coord_flip()

p3 <-ggarrange(pSSB_retro,pF_retro,pC_retro,pIAV_retro,pOFG_retro,pOFD_retro, 
               ncol=2, nrow=3, common.legend = TRUE, legend="bottom")
ggsave("Figure_3_orig.pdf",plot=p3,device="pdf",width=7,height=8)
}

# new version of Figure 3
{
  #quartz() # can only run this on Mac
  pSSB_retronew<-ggplot(data=filter(medretro_Fhist, metric=="avg_ssb_ssbmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=medval)) +
    geom_point((aes(colour = Retro_source, shape=F_history)))+
    theme_classic() + 
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(SSB_label) + ggtitle("A")+
    geom_hline(yintercept = 1) +
    scale_color_manual(values=point_cols)+
    coord_flip()
  
  pF_retronew<-ggplot(data=filter(medretro_Fhist, metric=="avg_f_fmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=medval)) +
    geom_point((aes(colour = Retro_source, shape=F_history)))+
    theme_classic() + 
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(F_label) + ggtitle("B")+
    geom_hline(yintercept = 1) +
    scale_color_manual(values=point_cols)+
    coord_flip()
  
  pC_retronew<-ggplot(data=filter(medretro_Fhist, metric=="avg_catch_msy"), aes(x=factor(IBMlab,level=IBM_order_new), y=medval)) +
    geom_point((aes(colour = Retro_source, shape=F_history)))+
    theme_classic() + 
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(C_label) + ggtitle("C")+
    geom_hline(yintercept = 1) +
    scale_color_manual(values=point_cols)+
    coord_flip()
  
  pIAV_retronew<-ggplot(data=filter(medretro_Fhist, metric=="iav_catch"), aes(x=factor(IBMlab,level=IBM_order_new), y=medval)) +
    geom_point((aes(colour = Retro_source, shape=F_history)))+
    theme_classic() + 
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(IAV_label) + ggtitle("D")+
    scale_color_manual(values=point_cols)+
    coord_flip()
  
  nING_retronew<-ggplot(data=filter(medretro_Fhist, metric=="n_gr_fmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=medval)) +
    geom_point((aes(colour = Retro_source, shape=F_history)))+
    theme_classic() + 
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(NING_label) + ggtitle("E")+
    scale_color_manual(values=point_cols)+
    coord_flip()
  
  nED_retronew<-ggplot(data=filter(medretro_Fhist, metric=="n_less_05_bmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=medval)) +
    geom_point((aes(colour = Retro_source, shape=F_history)))+
    theme_classic() + 
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(NED_label) + ggtitle("F")+
    scale_color_manual(values=point_cols)+
    coord_flip()
  
  p3n <-ggarrange(pSSB_retronew,pF_retronew,pC_retronew,pIAV_retronew,nING_retronew,nED_retronew, 
                 ncol=2, nrow=3, common.legend = TRUE, legend="bottom")
  ggsave("Figure_3_new.pdf",plot=p3n,device="pdf",width=7,height=8)
}

# Figure 4 ---------------------------------------------------------------
# F / Fmsy by Cmult, and  Short vs Long
# original
{
  pF_time_Fhist<-ggplot(data=time_Fhist, aes(x=factor(IBM,level=IBM_order), y=F)) +
    geom_point((aes(colour = Period, shape=F_history)))+
    theme_classic() + 
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(F_label) + ggtitle("A")+
    geom_hline(yintercept = 1) +
    scale_color_manual(values=point_cols)+
    coord_flip()
  
  pF_Cmult_Fhist<-ggplot(data=Cmult_Fhist, aes(x=factor(IBM,level=IBM_order), y=F)) +
    geom_point((aes(colour = Catch_mult, shape=F_history)))+
    theme_classic() + 
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(F_label) + ggtitle("B")+
    geom_hline(yintercept = 1) +
    scale_color_manual(values=point_cols)+
    coord_flip()

  p4 <- grid.arrange(pF_time_Fhist,pF_Cmult_Fhist,nrow=2)
  ggsave("Figure_4_orig.pdf",plot=p4,device="pdf",width=6,height=6)  
}

# new Figure 4
{
  pF_time_Fhistnew<-ggplot(data=filter(medtime_Fhist, metric=="avg_f_fmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=medval)) +
    geom_point((aes(colour = Period, shape=F_history)))+
    theme_classic() + 
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(F_label) + ggtitle("A")+
    geom_hline(yintercept = 1) +
    scale_color_manual(values=point_cols)+
    coord_flip()
  
  pF_Cmult_Fhistnew<-ggplot(data=filter(medCmult_Fhist, metric=="avg_f_fmsy"), aes(x=factor(IBMlab,level=IBM_order_new), y=medval)) +
    geom_point((aes(colour = Catch_mult, shape=F_history)))+
    theme_classic() + 
    theme(text=element_text(family="Times"))+
    xlab("") + ylab(F_label) + ggtitle("B")+
    geom_hline(yintercept = 1) +
    scale_color_manual(values=point_cols)+
    coord_flip()
  
  p4new <- grid.arrange(pF_time_Fhistnew,pF_Cmult_Fhistnew,nrow=2)
  ggsave("Figure_4_new.pdf",plot=p4new,device="pdf",width=6,height=6)  
}

# Figure 5 --------------------------------------------------------------- 
# original
# Long term C/MSY vs SSB/SSBmsy by realization and IBM
# uses scenario CF1A (see make_MS_tables_figures.R at bottom)
{
fig5plot <- ggplot(sims5, aes(x=l_avg_ssb_ssbmsy, y=l_avg_catch_msy)) +
  geom_point() +
  geom_vline(xintercept = 1, color="red", linetype="dashed") +
  geom_hline(yintercept = 1, color="red", linetype="dashed") +
  facet_wrap(~factor(IBMlab, level=rev(IBM_order))) +
  theme_classic() + 
  theme(text = element_text(family = "Times")) +
  labs(x=SSB_label, y=C_label)
ggsave(filename = "Figure_5_orig.pdf", plot=fig5plot, device="pdf", width=6,height=6)
}

# new Figure 5
{
  fig5new <- ggplot(simsfig5, aes(x=avg_ssb_ssbmsy, y=avg_catch_msy)) +
    geom_point(alpha=0.15) +
    geom_vline(xintercept = 1, color="red", linetype="dashed") +
    geom_hline(yintercept = 1, color="red", linetype="dashed") +
    facet_wrap(~factor(IBMlab, level=rev(IBM_order_new))) +
    theme_classic() + 
    theme(text = element_text(family = "Times")) +
    labs(x=SSB_label, y=C_label)
  ggsave(filename = "Figure_5_new.pdf", plot=fig5new, device="pdf", width=6,height=6)
}

# Figure 5 alternative 1 (F/Fmsy vs SSB/SSBmsy)
{
  fig5alt1 <- ggplot(simsfig5, aes(x=avg_ssb_ssbmsy, y=avg_f_fmsy)) +
    geom_point(alpha=0.15) +
    geom_vline(xintercept = 1, color="red", linetype="dashed") +
    geom_hline(yintercept = 1, color="red", linetype="dashed") +
    facet_wrap(~factor(IBMlab, level=rev(IBM_order_new))) +
    theme_classic() + 
    theme(text = element_text(family = "Times")) +
    labs(x=SSB_label, y=F_label)
  ggsave(filename = "Figure_5_alt1.pdf", plot=fig5alt1, device="pdf", width=6,height=6)
}

# Figure 5 alternative 2 (Catch/MSY vs F/Fmsy)
{
  fig5alt2 <- ggplot(simsfig5, aes(x=avg_f_fmsy, y=avg_catch_msy)) +
    geom_point() +
    geom_vline(xintercept = 1, color="red", linetype="dashed") +
    geom_hline(yintercept = 1, color="red", linetype="dashed") +
    facet_wrap(~factor(IBMlab, level=rev(IBM_order_new))) +
    theme_classic() + 
    theme(text = element_text(family = "Times")) +
    labs(x=F_label, y=C_label)
  ggsave(filename = "Figure_5_alt2.pdf", plot=fig5alt2, device="pdf", width=6,height=6)
}
