==> Manuscript/MS_tables_figs_rev1/4_make_CJFAS_SupplMat_pdf.Rmd <==
---
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
```

\newpage

# Supplemental Materials for Data Rich but Model Resistant: An Evaluation of Data-Limited Methods to Manage Fisheries with Failed Age-based Stock Assessments

  
Christopher M. Legault^1^, John Wiedenmann^2^ Jonathan J. Deroba^1^, Gavin Fay^3^, Timothy J. Miller^1^, Elizabeth N. Brooks^1^, Richard J. Bell^4^, Joseph A. Langan^5^, Jamie M. Cournane^6^, Andrew W. Jones^1^, Brandon Muffley^7^


^1^National Marine Fisheries Service, Northeast Fisheries Science Center, Woods Hole, MA, USA  
^2^Department of Ecology, Evolution and Natural Resources, Rutgers, The State University of New Jersey, New Brunswick, New Jersey, USA  
^3^University of Massachusetts Dartmouth, School for Marine Science
and Technology, Massachusetts, USA  
^4^The Nature Conservancy, Rhode Island, USA  
^5^University of Rhode Island, Graduate School of Oceanography, Narragansett, RI, USA  
^6^New England Fishery Management Council, Newburyport, MA, USA  
^7^Mid-Atlantic Fishery Management Council, Dover, DE, USA  


We provide additional details of the simulation study (Figures S1-S4) and results (Figures S5-S37) for readers seeking more granular results of the 16 scenarios. The distribution of Mohn's rho values is shown in Figure S1, illustrating that we achieved a median rho of 0.5 for SSB. The difference in true $F_{MSY}$ relative to the fishing history scenarios is shown in Figure S2; this provides some insight into the choice for which reference point was used for calculating overfishing. A similar plot for spawning stock biomass in shown in Figure S3 demonstrating the starting conditions for the feedback portion of the simulations. Retrospective patterns continued in the feedback portion of the closed-loop simulations, as seen in two realizations for both sources of the retrospective pattern in Figure S4. An alternative way of viewing the median results from Figure 1 in a more summarized approach that does not include uncertainty from the 1000 realizations is seen in the plot of the first two principal components of Figure S5. Plots similar to Figure 5 are seen in Figures S6-S21, showing the catch and spawning stock biomass relative to their MSY reference points for each of the 1000 realizations with each figure representing a different scenario. Figures S22-S37 are similar to Figures S6-S21 except that the y-axis is now fishing mortality relative to $F_{MSY}$, making them similar to Kobe plots. 


```{r retrores, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"../../demonstrations/tim","retro_results_plots.png"))
```

Figure S1. Distribution of 50 Mohn’s rho values for no misspecification (top row), catch misspecification (middle row), and natural mortality misspecification (bottom row) for spawning stock biomass (SSB), fishing mortality rate (Fbar), and recruitment (R). Panels are defined by retrospective type (none, catch, natural mortality), fishing intensity (1=always overfishing, 2=overfishing reducing to $F_{MSY}$), and number of selectivity blocks (1 or 2).
\newpage

```{r F, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"../../demonstrations/chris/base_sims","F.png"))
```

Figure S2. Fishing mortality rate (F) during the base period (black line) for the 8 combinations of F history (F = $F_{MSY}$ in second half of base period, O = overfishing throughout), number of fishery selectivity blocks (1 or 2), and retrospective source (catch or natural mortality, M). The horizontal blue line denotes the F at maximum sustainable yield ($F_{MSY}$) computed from conditions in the first year of the base period (1970) while the red dashed horizontal line denotes the $F_{MSY}$ computed from conditions in the terminal year in the base period (2019).
\newpage

```{r SSB, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"../../demonstrations/chris/base_sims","SSB.png"))
```

Figure S3. Spawning stock biomass (SSB) during the base period (black line) for the 8 combinations of F history (F = $F_{MSY}$ in second half of base period, O = overfishing throughout), number of fishery selectivity blocks (1 or 2), and retrospective source (catch or natural mortality, M). The horizontal blue line denotes the F at maximum sustainable yield ($F_{MSY}$) computed from conditions in the first year of the base period (1970) while the red dashed horizontal line denotes the $F_{MSY}$ computed from conditions in the terminal year in the base period (2019).
\newpage

```{r retrofeed, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"../../demonstrations/tim","example_feedback_SSB_retro.png"))
```

Figure S4. Mohn's rho for spawning stock biomass calculated from the statistical catch at age model during the feedback period of the simulations for two realizations (columns) and the two sources of retrospective (catch and natural mortality).
\newpage

```{r pcaplot, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat","pca.png"))
```

Figure S5. Principal components analysis of the median values of the metrics shown in Figure 1.
\newpage


```{r loop, echo=FALSE}
myfile <- list()
mytext <- list()
myfile2 <- list()
mytext2 <- list()
mycount <- 0
for (i in 1:2){
  s1 <- ifelse(i==1,"C","M")
  l1 <- ifelse(i==1,"catch","natural mortality")
  for (j in 1:2){
    s2 <- ifelse(j==1,"F","O")
    l2 <- ifelse(j==1,"reduced to","remaining at 2.5 times")
    for (k in 1:2){
      s3 <- k
      l3 <- ifelse(k==1,"was a single selectivity block","were two selectivity blocks")
      for (m in 1:2){
        s4 <- ifelse(m==1,"A","R")
        l4a <- ifelse(m==1,"no","a")
        l4b <- ifelse(m==1,"1","0.75")
        
        mycount <- mycount + 1
        myfile[[mycount]] <- paste0(s1,s2,s3,s4,"_C_vs_SSB.png")
        myfile2[[mycount]] <- paste0(s1,s2,s3,s4,"_F_vs_SSB.png")
        mytext[[mycount]] <- paste0("\nFigure S", mycount+5, ". Relationship between long-term average catch and spawning stock biomass relative to their reference points by method. Each point represents the average for years 21-40 in the feedback period for a single iteration of a scenario. The scenario shown is where ", l1, " was the source of the retrospective pattern with $F$ ", l2, " $F_{MSY}$ in the second half of the base period, there ", l3, ", and where ", l4a, " buffer was applied to the catch advice (catch multiplier = ", l4b, ").")
        mytext2[[mycount]] <- paste0("\nFigure S", mycount+5+16,". Relationship between long-term average fishing mortality rate and spawning stock biomass relative to their reference points by method. Each point represents the average for years 21-40 in the feedback period for a single iteration of a scenario. The scenario shown is where ", l1, " was the source of the retrospective pattern with $F$ ", l2, " $F_{MSY}$ in the second half of the base period, there ", l3, ", and where ", l4a, " buffer was applied to the catch advice (catch multiplier = ", l4b, ").")
      }
    }
  }
}
```

```{r p1, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile[[1]]))
```
`r mytext[[1]]`
\newpage

```{r p2, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile[[2]]))
```
`r mytext[[2]]`
\newpage

```{r p3, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile[[3]]))
```
`r mytext[[3]]`
\newpage

```{r p4, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile[[4]]))
```
`r mytext[[4]]`
\newpage

```{r p5, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile[[5]]))
```
`r mytext[[5]]`
\newpage

```{r p6, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile[[6]]))
```
`r mytext[[6]]`
\newpage

```{r p7, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile[[7]]))
```
`r mytext[[7]]`
\newpage

```{r p8, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile[[8]]))
```
`r mytext[[8]]`
\newpage

```{r p9, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile[[9]]))
```
`r mytext[[9]]`
\newpage

```{r p10, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile[[10]]))
```
`r mytext[[10]]`
\newpage

```{r p11, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile[[11]]))
```
`r mytext[[11]]`
\newpage

```{r p12, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile[[12]]))
```
`r mytext[[12]]`
\newpage

```{r p13, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile[[13]]))
```
`r mytext[[13]]`
\newpage

```{r p14, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile[[14]]))
```
`r mytext[[14]]`
\newpage

```{r p15, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile[[15]]))
```
`r mytext[[15]]`
\newpage

```{r p16, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile[[16]]))
```
`r mytext[[16]]`
\newpage

```{r p21, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile2[[1]]))
```
`r mytext2[[1]]`
\newpage

```{r p22, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile2[[2]]))
```
`r mytext2[[2]]`
\newpage

```{r p23, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile2[[3]]))
```
`r mytext2[[3]]`
\newpage

```{r p24, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile2[[4]]))
```
`r mytext2[[4]]`
\newpage

```{r p25, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile2[[5]]))
```
`r mytext2[[5]]`
\newpage

```{r p26, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile2[[6]]))
```
`r mytext2[[6]]`
\newpage

```{r p27, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile2[[7]]))
```
`r mytext2[[7]]`
\newpage

```{r p28, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile2[[8]]))
```
`r mytext2[[8]]`
\newpage

```{r p29, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile2[[9]]))
```
`r mytext2[[9]]`
\newpage

```{r p210, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile2[[10]]))
```
`r mytext2[[10]]`
\newpage

```{r p211, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile2[[11]]))
```
`r mytext2[[11]]`
\newpage

```{r p212, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile2[[12]]))
```
`r mytext2[[12]]`
\newpage

```{r p213, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile2[[13]]))
```
`r mytext2[[13]]`
\newpage

```{r p214, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile2[[14]]))
```
`r mytext2[[14]]`
\newpage

```{r p215, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile2[[15]]))
```
`r mytext2[[15]]`
\newpage

```{r p216, echo=FALSE, message=FALSE, warnings=FALSE, out.width='7in'}
knitr::include_graphics(file.path(getwd(),"supplmat",myfile2[[16]]))
```
`r mytext2[[16]]`
\newpage
